# [rush] Optimize the execution speed of Rush