--- a/common/changes/@microsoft/rush/main_2024-11-18-08-13.json
+++ b/common/changes/@microsoft/rush/main_2024-11-18-08-13.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@microsoft/rush",
+      "comment": "Optimize the execution speed of Rush",
+      "type": "none"
+    }
+  ],
+  "packageName": "@microsoft/rush"
+}