--- a/common/changes/@microsoft/rush/main_2024-11-18-08-13.json
+++ b/common/changes/@microsoft/rush/main_2024-11-18-08-13.json
@@ -2,7 +2,7 @@
   "changes": [
     {
       "packageName": "@microsoft/rush",
-      "comment": "Optimize the execution speed of Rush",
+      "comment": "Add a new experiment flag `enableSubpathScan` that, when invoking phased script commands with project selection parameters, such as `--to` or `--from`, only hashes files that are needed to compute the cache ids for the selected projects.",
       "type": "none"
     }
   ],