--- a/common/changes/@rushstack/package-deps-hash/main_2024-11-18-08-13.json
+++ b/common/changes/@rushstack/package-deps-hash/main_2024-11-18-08-13.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@rushstack/package-deps-hash",
+      "comment": "Optimize the execution speed of Rush",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@rushstack/package-deps-hash"
+}