--- a/common/changes/@rushstack/package-deps-hash/main_2024-11-18-08-13.json
+++ b/common/changes/@rushstack/package-deps-hash/main_2024-11-18-08-13.json
@@ -2,8 +2,8 @@
   "changes": [
     {
       "packageName": "@rushstack/package-deps-hash",
-      "comment": "Optimize the execution speed of Rush",
-      "type": "patch"
+      "comment": "Add a new optional parameter `filterPath` to `getRepoStateAsync` that limits the scope of the git query to only the specified subpaths. This can significantly improve the performance of the function when only part of the full repo data is necessary.",
+      "type": "minor"
     }
   ],
   "packageName": "@rushstack/package-deps-hash"