--- a/libraries/rush-lib/assets/rush-init/common/config/rush/experiments.json
+++ b/libraries/rush-lib/assets/rush-init/common/config/rush/experiments.json
@@ -102,5 +102,11 @@
   /**
    * When using cobuilds, this experiment allows uncacheable operations to benefit from cobuild orchestration without using the build cache.
    */
-   /*[LINE "HYPOTHETICAL"]*/ "allowCobuildWithoutCache": true
+   /*[LINE "HYPOTHETICAL"]*/ "allowCobuildWithoutCache": true,
+
+   /**
+   * By default, rush perform a full scan of the entire repository. For example, Rush runs `git status` to check for local file changes.
+   * When this toggle is enabled, Rush will only scan specific paths, significantly speeding up Git operations.
+   */
+  /*[LINE "HYPOTHETICAL"]*/ "enableSubpathScan": true
 }