@dmichon-msft 2024-11-18
Conceptually this seems fine, but I'd want to do a lot of stress testing and verify the interaction with, e.g. the `getAdditionalFilesFromRushProjectConfigurationAsync` function. Adding the filter will cause non-project files to no longer be included in the initial hash set, which means they will require an additional round trip to get the hashes for, despite being included in the git index.

The other problem, though, is that the raw response of this call is potentially consumed by plugins that make use of the `inputSnapshot` object, so if it is suddenly missing large swathes of files that would be unexpected.

I'm reasonably confident that the performance impact of filtering `git ls-tree` is negligible; in my experience most of the overhead comes from the `git status` call.