@L-Qun 2024-11-18
> During my typical testing:
> 
> ```
> time git ls-tree -z -r --full-name HEAD -- > /dev/null
> 
> real    0m0.100s
> user    0m0.087s
> sys     0m0.013s
> ```
> 
> ```
> time git status -z -u --no-renames --ignore-submodules --no-ahead-behind -- > /
> dev/null
> 
> real    0m0.648s
> user    0m0.238s
> sys     0m0.787s
> ```
> 
> The issue is that the call to `git status` scans for untracked files with changes, and that operation is expensive. If we definitively _know_ that there are no unstaged changes, we could speed up evaluation by only dumping the files that are staged, but in typical local development we can't guarantee that.
> 
> I'd love to be able to reduce the scope, but it would be a breaking change, because we expose the full list of tracked files and their hashes here:
> 
> https://github.com/microsoft/rushstack/blob/da48ac39391567a759524f2e614abb4cfe9687ed/libraries/rush-lib/src/logic/incremental/InputsSnapshot.ts#L127

Yes, It would be a breaking change. I modified this comments.